requirements.txt
# Core App Logic
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.3
starlette==0.37.2

# Async + File Upload Support
aiofiles==23.2.1
python-multipart==0.0.9

# Image and Media Handling
Pillow==10.3.0

# HTTP Requests
requests==2.32.3

# Environment Variables
python-dotenv==1.0.1

# LLM & Generative AI
litellm==1.41.17
google-generativeai

# External Tools
gradio_client==0.15.0

# Optional Frontend
streamlit
crewai
